+++ example/lib/logic/event_state.dart
@@ -1,12 +1,12 @@
 import 'package:equatable/equatable.dart';
-import 'package:easy_calendar/easy_calendar.dart';
+import 'package:eventide/eventide.dart';
 import 'package:value_state/value_state.dart';
 
 typedef EventState = Value<EventValue>;
 
 class EventValue extends Equatable {
-  final ECCalendar calendar;
-  final List<ECEvent> events;
+  final ETCalendar calendar;
+  final List<ETEvent> events;
 
   @override
   List<Object?> get props => [calendar, events];