+++ ios/Classes/EasyEventStore/EasyEventStore.swift
@@ -1,6 +1,6 @@
 //
 //  EasyEventStore.swift
-//  easy_calendar
+//  eventide
 //
 //  Created by CHOUPAULT Alexis on 31/12/2024.
 //