+++ ios/Classes/EventidePlugin.swift
@@ -1,8 +1,8 @@
 import Flutter
 import UIKit
 
-public final class EasyCalendarPlugin: NSObject, FlutterPlugin {
-    
+public class EventidePlugin: NSObject, FlutterPlugin {
+  
     public static func register(with registrar: FlutterPluginRegistrar) {
         let easyEventStore = EasyEventStore(eventStore: EKEventStoreSingleton.shared.eventStore)
         let permissionHandler = PermissionHandler(eventStore: EKEventStoreSingleton.shared.eventStore)