+++ lib/eventide.dart
@@ -1,0 +1,5 @@
+library eventide;
+
+export 'src/eventide.dart' show Eventide;
+export 'src/eventide_platform_interface.dart' show ETCalendar, ETEvent;
+export 'src/eventide_exception.dart' show ETException, ETPermissionException, ETNotEditableException, ETNotFoundException, ETGenericException;