+++ pigeons/calendar_api.dart
@@ -4,11 +4,11 @@ import 'package:pigeon/pigeon.dart';
 @ConfigurePigeon(PigeonOptions(
   dartOut: 'lib/src/calendar_api.g.dart',
   dartOptions: DartOptions(),
-  kotlinOut: 'android/src/main/kotlin/sncf/connect/tech/easy_calendar/CalendarApi.g.kt',
-  kotlinOptions: KotlinOptions(package: 'sncf.connect.tech.easy_calendar'),
+  kotlinOut: 'android/src/main/kotlin/sncf/connect/tech/eventide/CalendarApi.g.kt',
+  kotlinOptions: KotlinOptions(package: 'sncf.connect.tech.eventide'),
   swiftOut: 'ios/Classes/CalendarApi.g.swift',
   swiftOptions: SwiftOptions(),
-  dartPackageName: 'easy_calendar',
+  dartPackageName: 'eventide',
 ))
 
 @HostApi()