+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 0.1.0
 
-* TODO: Describe initial release.
+Initial release with features:
+* create/retrieve/delete calendars
+* create/retrieve/delete events
+* create/delete reminders
+* ask for system calendar permission (although the plugin handles it automatically)